# src/AspNetCore.SampleOpenApi.Api/AspNetCore.SampleOpenApi.Api_v1.0.json
  {
  "openapi": "3.1.1",
  "info": {
    "title": "OpenApi Sample API",
    "description": "Sample API to test migration from Swashbuckle to AspNetCore OpenApi",
    "version": "1.0"
  },
  "paths": {
    "/api/v1.0/errordemo/default": {
      "get": {
        "tags": [
          "ErrorDemo"
        ],
        "operationId": "GetDefaultError",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1.0/errordemo/problem": {
      "get": {
        "tags": [
          "ErrorDemo"
        ],
        "operationId": "GetProblemDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1.0/errordemo/validationproblem": {
      "get": {
        "tags": [
          "ErrorDemo"
        ],
        "operationId": "GetValidationProblemDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1.0/forecasts": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForcasts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Person": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Status": {
        "enum": [
          "Active",
          "Inactive"
        ],
        "type": "string"
      },
      "WeatherForecast": {
        "required": [
          "id",
          "date",
          "temperatureC",
          "reportedBy",
          "summary",
          "currStatus",
          "createdBy",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "temperatureC": {
            "maximum": 5000,
            "minimum": -274,
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int64"
          },
          "reportedBy": {
            "type": "string",
            "format": "email"
          },
          "summary": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "prevStatus": {
            "$ref": "#/components/schemas/Status"
          },
          "currStatus": {
            "$ref": "#/components/schemas/Status"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Person"
          },
          "modifiedBy": {
            "$ref": "#/components/schemas/Person"
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "ErrorDemo"
    },
    {
      "name": "WeatherForecast"
    }
  ]
}